Major Update
The simulation can now be stopped an manually stepped through.
Added a restart button - so the simulation can be restarted without having to close and reopen the program.

Added "Gossip Details" which shows Info about a selected gossip as well as all the event that have occurred to the gossip

NPC details now show their personalities as well as relationships are now displayed in a table structure.

The camera can no longer be zoomed in and out.

Updated README to reflect new changes